Fix issue of white space in sitelinkgroupview heading
This is an alternative proposal to I5fb924e.

Note that this changes the behavior of the heading when the headline
in front of the edit buttons contains a lot of text. Before (when it
was a table): The two elements were placed in two table cells. The text
in each cell wrapped to two short lines, headline left, buttons right.
After (with inline-block): The buttons wrap to a new line, under the
headline.

Bug: T92431
Change-Id: I92aa716e2545103b7a7f3a6dcbb684d6d856cd2b