Run UpdateRepo jobs on time() + 2, not time() + 1
Just adding +1 to time() doesn't mean we wait at least
a second, it means we'll wait max. a second.

Due to this actual delay depends on the time until
the next second... due to this add 2, so that we wait at
least one second which should be enough for slaves to
catch up in case of no replag.

Bug: T92789
Change-Id: Ieb2028d41f1e9edadbab6ee7c175514c9458eb73