Update patch set 1
Patch Set 1:

I tried this merged with current master. On master, the edit button does not overlap with the table header. With this change, it does. I don't know how to get out of this, but to me it looks really bad that way; even though I generally like this change.

Patch-set: 1