Update patch set 1
Patch Set 1:

fortunately we have $wgWBSharedSettings['useLegacyChangesSubscription'] = true; on wikidata, so this is not an issue for us.

but I had this wrong on my own dev wiki and the setting in the client didn't take effect. :(  clear documentation on this sort of thing is essential.

Patch-set: 1
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0