Update patch set 2
Patch Set 2: Code-Review-1

Please check my comment. Beside that, looks fine to me as soon as the patch this one is depending on is merged. Otherwise only half of it makes sense so far. (Even though the second half has to be updated in the way you did it either way.)

Patch-set: 2
Label: Code-Review=-1