More clear variable and method naming in ChangesSubscriptionTableBuilder
ChangesSubscriptionTableBuilder seems to have nothing to do
with EntityPerPage, among other things that were not as clear
as they could be.

the slurpSubscriptions method is tricky to name because it
does more than one thing for each row it processes, so couldn't
come up with the perfect name but think this is better.

Change-Id: I489188c7dba5cb871f15b73292b44c9ce9dba158