Update patch set 6
Patch Set 6:

Now I changed and tested it on the given test wiki and it match perfectly to what I am want. Feel free to ask why, this is not getting complicated for no reason. Empty description should match user language and direction not the language entry. It should applied every time because when user adds or clears an description or alias, a new and correspond language and direction shoul be applied and it should remains with its previous state.

Patch-set: 6
Label: Verified=0