Update patch set 4
Patch Set 4:

Having different classes for different kinds of dumps makes sense to me. Reducing code complexity makes it easier to add new features, and to add and avoid bugs. 

Adding more to the pile is quicker now, but will cost time in the long run. 

As I said, we don't have to block this patch on that. But I do want to have some kind of commitment to untangling the code soon. And usually, if it's not blocking a feature, it will not happen. So it's best to invest the time to refactor when wanting a feature, because otherwise, adding features is going to get more and more tricky.

Patch-set: 4