Update patch set 1
Patch Set 1:

Is it really necessary to have all these if/then/else switches all over the code? Isn't there a compat wrapper in core? Can't we have a method in the base class for getting and using that?

Also note that Wikibase is not kept compatible with old versions of mediawiki. We move along pretty much in sync with core. If you want to use Wikibase with an old version of MW, use an old version of WB.

Patch-set: 1