Update patch set 4
Patch Set 4:

The current implementation is a single purpose internal cleanup function. But what I see in it is really a generic data model converter interface. Which we will need anyway, for use in several places (depending on the outcome of the current discussion, for rendering, or for inclusion in JSON dumps, etc).

Patch-set: 4