Update patch set 4
Patch Set 4:

It's not different dumps, it's the same dump with different options enabled. When you have a white car and a yellow car, you don't create two classes just because color is different. Making each method and half-method into class does not make sense to me, especially when the things are changing a lot. Reading this code and navigating it and changing it would be very hard and performance would suffer too because of constant indirection. I think we should keep things simple and have one class do one thing - if the thing is producing RDF or an entity, then this class should do it, no need to have separate class for producing RDF of each detail. We won't need anything but RDF of whole entity, ever, so that's what the class should be doing.

Patch-set: 4
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0