Update patch set 20
Patch Set 20:

This is without the performance tweaks, just checking out this branch. It is a big step forward compared to EasyRDF but we can do even better :)

I'm ok with merging this one anytime you want. There's another direction to tweak - we may want to fix how prefixes are done and eliminate checkPrefixes from dump. Now that we control the output it should not be needed anymore.

Patch-set: 20