Update patch set 5
Patch Set 5: Code-Review+2

This should have been considered as a deployment blocker because the edit-summary of that special page is completely wrong. Let's not block this fix on another patch where it is unclear when it gets merged. I've created bug T93860 to add tests once the testing framework for special pages is merged.

This patch does NOT fix the problem that it is not possible to delete labels, descriptions or all aliases using the special page. This issue is tracked in bug T91387 and resolved in https://gerrit.wikimedia.org/r/#/c/194105/.

Patch-set: 5
Reviewer: Tobias Gritschacher <170@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2