Update patch set 4
Patch Set 4:

The more important question is still unanswered. I find this code very suspicious. This is manipulating the request under test in code that's not part of the test. How do we know all tests making use of POST requests are testing edits and need an edit token? Will the presence of an unexpected token break tests? Why is the token not set via a method that's only called when the token is really needed?

If there is no other way I will merge this. But not blindly, with no answer.

I also added a FIXME in my follow-up If7911ba.

Patch-set: 4