Fix focus on value field on property select
With I6e694dd and possibly other recent snakview changes, keydown
event now occurs after entityselectorselected. This meant that
_tabToValueView is not set yet when entityselectorselected
is executed and focus is never then set to the value input.

I suppose perhaps this worked before because this check and
the focus happened inside "self.options.entityStore.get( entityId ).done"
so the timing worked.

This allows focus to be set on tab or enter keydown or select.

Bug: T93864
Change-Id: Id386edf0942f2378c11e24528a2957316099c91b