Update patch set 5
Patch Set 5:

(1 comment)

I don't understand the n-triples => ntriples move. 
Also, it's unrelated to the subdocument issue and should be in a separate patch. Should be easy to split, since it affects different files.

I think the subdocument state issue could be solved more elegantly, but this is fine solution seems good enough. That aspect of the whole state transition system is a bit messy, I ended up changing it around a few times.

Patch-set: 5