Update patch set 9
Patch Set 9:

Thiemo, please see latest patch, some of the algorithm was changed. The basic premise though is that for Julian, we can not convert "weird" dates to gregorian, so conversion will fail and they'll be ignored (meaning, represented as string). For Gregorian, we need to make it real valid gregorian dates. For dates far in the past, it means ignoring month and day, because we can not validate them, and if they'd be invalid the loading app would either fail to load data completely or fail to load it as a valid date, which will make it unsearchable. It is better to have less precise but searchable data than none at all.

If you need the original data you have it in the deep value which is always kept as is. Cleaned value is converted value for search purposes and as such should be fit for this purpose.

Patch-set: 9