Factor database access out of WikiPageEntityRevisionLookup
Introduces WikiPageEntityMetaDataLookup which is able to retrieve
the meta data that WikiPageEntityRevisionLookup needs for both accessing
entities and to verify that an entity revision from cache is still
the latest version.

WikiPageEntityMetaDataLookup also supports getting meta data for
multiple entities at once, but we're not yet making use of that.
That will be used for implementing T87238.

Change-Id: Ie7f76921e1d1ffd2e67c33c597e9116f2148ed3d