Update patch set 1
Patch Set 1:

I still think we should have the "major" version ID in the schema URI (i.e. 0.1, 0.2 for beta, 1, 2, 3 etc for releases). If we don't, we basically commit to never making a breaking change to the ontology. Making a breaking change without changing concept URIs is bound to create a major mess. The XSD date issue seems to be a good (or rather bad) example of how this can go wrong.

Patch-set: 1
Label: Code-Review=0