Use wgPageName to construct an mw.Title in linkItem init
Right now we use wgTitle (which strips of the namespace) and pass
a default namespace to use to mw.Title. That works, unless wgTitle
starts with a string that is a valid namespace itself (such as
"Wikipedia:…" on the page "Kategorie:Wikipedia:…"), as mw.Title
uses a namespace in the given title text over the default.

Bug: T94119
Change-Id: I5af216fb606f2671cb98bfaf7b6e94c29161c9be