Update patch set 2
Patch Set 2:

For the same reason ISO disallows years with less than four digits: the timestamp 01-02-03 is ambiguous. THis problem isn't as big in the API as it would be when parsing user input, but still. We require two digits for the day, right? so why not require four for the year?

Patch-set: 2