Introduce EntityIdPlainLinkFormatter for SummaryFormatter
The wikitext links that SummaryFormatter outputs must not
include a display title, as that will stop the onLinkBegin
hook handler from altering the link (as it shouldn't magically
add display titles to links that already have one).

In order to fix this, I created a new EntityIdPlainLinkFormatter
which is basically the pre fb60de9acf5ffa5347fc0f57656073154e170c97
EntityIdLinkFormatter for use in case we don't want a display title.

Bug: T93804
Change-Id: Ib362ae7be39f94bd084665a58524c3c40bce99b8