Drop .coveralls.yml
I think this service is not used any more and this file is an unused,
possibly outdated (look how it references the "lib" directory and
nothing else) fragment now.

I may be wrong. If you think I am, please show me the place where
Coveralls is used and abandone this patch.

Change-Id: I3e1a1428b7f4b138ae9f5371215917c39bc6b15e