Update patch set 2
Patch Set 2:

> For the same reason ISO disallows years with less than four digits

Sure, because ISO allows timestamps with no sign and no time part. Sure, "01-02-03" is ambiguous. But this example does not apply to this validator. Nothing is optional here. The sign character is non-optional. The "T", hour, minute, second and trailing "Z" are non-optional. There is no way anybody can confuse "+01-02-03T00:00:00Z". Really, what else could it mean?

Look at how I solved exactly this in https://github.com/DataValues/Time/pull/27 (which is stuck in limbo for more than two months now, by the way).

> why not require four for the year?

Sorry? Because we want to support 1, 2 and 3-digit years? Because we decided to not enforce padding on the year? The 4-digit padding we talked about is for convenience. I really don't see a reason why we should ever *REJECT* years with less than 4 digits. This was never done in Wikibase related code and never part of the discussions we had.

-2 because this needs discussion. In person. Personally I think this is a misunderstanding.

Patch-set: 2