Update patch set 2
Patch Set 2:

A diff has and old and a new ›side‹, but these ›sides‹, but a ›side‹ does not have to have a snak. The oldest snak is the oldest snak in the diff, which is the old snak, if it exists, otherwise the new snak. The newest snak is the new snak, if it exists, otherwise the old snak.

Patch-set: 2