Update patch set 12
Patch Set 12: Code-Review-1

The code looks good, but it assumes that Julian dates are stored unconverted in entity data. This isn't the case, and the jury is still out on the question whether we want it to be the case.

I'd be fine with merging this if julian conversion could be turned on and off using define( 'WIKIBASE_STORE_JULIAN' ) or some such.

Patch-set: 12
Label: Code-Review=-1