Update patch set 2
Patch Set 2:

I think it's not the job of this validation step to apply the convenience-padding we talked about. https://github.com/DataValues/Time/pull/33 and https://github.com/DataValues/Time/pull/56 are the relevant patches. And even if we want to add this to the set of validators just to be sure, this would break production as it is now because not all parsers are guaranteed to return years that are padded to 4 digits. This requires a closer look at all parsers, more test cases in all parser tests and a new release of the time component.

In addition to that, the regex currently in this patch is not sufficient because years like 02015 should be disallowed.

Patch-set: 2