Update patch set 1
Patch Set 1: Code-Review-1

I think this was done on purpose. Labels and descriptions can be 250 characters long. So can aliases. Imagine a box with two or more languages, each containing two very long aliases. This box will already span the whole screen, and it's not even an extreme example.

I suggest to keep the CSS as it is for aliases.

I suggest adding a max-height or something like that.

Also note that this patch does not fix the problem that a label input field is still a single line instead of a textarea. This is confusing: Before editing the label is multi-line, but when clicking edit it becomes single-line. That's the main reason for my -1.

I'm not saying this is a bad patch. It's definitely going into the right direction. Thanks a lot for poking at the code!

Patch-set: 1
Label: Code-Review=-1