Update patch set 12
Patch Set 12:

ok, I can do the switch. The problem is the tests - if you change the switch, you'll have to change the tests. Also, if dates are Gregorian we may as well have deep value as non-string, since it's close enough to what we need. Right now we use string to preserve the julian original, but if there's no julian original there's no point in it.

Patch-set: 12
Label: Verified=0