Remove keydown event from property selector input
As far as I can tell, keydown event is already handled in
jquery.ui.suggester (_attachInputEventHandlers) and always
triggers entityselectorselected.

We can remove the event handling here and the property
still gets selected and focus is appropriately set in
the value input box for value snaks.

Having keydown event *and* entityselectorselected sometimes
results in the value field getting stuck with a spinner,
depending on timing of events. This change helps to avoid
that happening.

Bug: T96466
Change-Id: If192c130a53cb2a9e7a9add103e356c60d2459c2