Hook up subscription lookup based on settings.
This allows sitelinks based subscription tracking, explicit subscription tracking,
as well as a hybrid mode, to be configured.

DEPLOY: if useLegacyChangesSubscription=true is set on the repo,
now set subscriptionLookupMode='sitelinks'. useLegacyChangesSubscription 
remains in use for clients, and could be kept for the repo, in case we end
up rolling back the code.

Bug: T66590
Bug: T90755
Change-Id: Ie1f1c372c1f5c8ff9db6b13dcbbadeee5aaf273f