[DNM] Drop unused EntityTestHelper::getTestEntity() method
I found this while searching for deprecated DataModel methods.
Some of this code is actually broken and should cause serious errors.
I was wondering why this does not happen. Reason is: this is never
executed.

This patch actually reverts Ifdd5b29! It's *newer* code that was
introduced to replace *older* code but was never used and is outdated
now. What to do? Fix it and start using it? Is this worth our time?

Bug: T78290
Change-Id: I928db375cd1c222b5d9fdedbbc89990fdefa2ff5