Allow wikibase-time-precision-* to be translated, expand Qqq
Although they're probably not terribly well covering all edge cases
in all the languages out there, having them translatable as they are
is still better than having them English only.

At some point we may come up with a better way to localize this,
but that's not going to happen soon, so we should rather have this
localized now.

Bug: T50962
Change-Id: I57bfb737b68010020b75a592e864f0298fe0046c