Fix revision id used for EntityTermsView placeholder
Requesting ParserOutput from EntityContent with null or 0
revision id happens during saving when revision id is not
known or assigned yet.

Trying to get latest from the title is not the right thing
to do here, and results in issues like outdated revision
used to display entity terms in the "other languages" box.

Bug: T90893
Change-Id: Ic45c727b6acdab58881d99460956bbe2e7c90a1c