Update patch set 7
Patch Set 7:

Why do you think this is a problem?

The "value object" aspect was super-super trivial. Basically an array of strings. I do not think such a class adds much value, which is why I'm phasing it out.

The "registry" aspect is now part of the factory. Isn't this how it's supposed to be? You create a factory, teach the factory what it should do and use it. I still wonder why these two aspects were split into two classes in the first place.

Could be that this needs a bite more refactoring to be sane, like renaming methods. Please add suggestions.

Patch-set: 7
Label: Verified=0