Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Hoo, what "change in semantics"?

I like that this is backwards-compatible. We could run a script to re-calculate all weights. But this is not necessary, in my opinion. Most entities do have a sane weight and the few that don't are edited quite often and get a magic update. What do you think? Is this ok?

Patch-set: 1
Label: Code-Review=+1