Update patch set 11
Patch Set 11: Code-Review+1

Wow. I had a look at the old 1000+ lines class. Sorry to say that, but this *must* be split. The split done in this patch looks sane. It can be improved much more, obviously, but shouldn't in this patch. I suggest to merge this as it is and continue refactoring in later patches.

Patch-set: 11
Label: Code-Review=+1