Update patch set 4
Patch Set 4:

EntityRevision objects that may or may not have their revision set to 0 seem very scary to me, we'd have to check all code that accesses the revision ID from an EntityRevision object.

I agree that using Title::getLatestRevisionId is probably not the right thing to do. But what would be the correct thing?

Patch-set: 4