Update patch set 4
Patch Set 4:

@daniel we don't know what the revision id is at this point. 

in this case, we are getting parser output via WikiPage->prepareContentForEdit.

when the save does happen, then we get the revision back as part of the Status object

$revision = $status->value['revision'];

and if no new revision was created, then $revision = $page->getRevision();

i don't seen a viable alternative here. EntityRevision has revision as optional and seems to be for a reason, such as this case.

Patch-set: 4