Update patch set 2
Patch Set 2:

(2 comments)

This should be fine, but the patch could be much smaller if it didn't insist to rename LabelLookup to LabelDescriptionLookup. With LabelDescriptionLookup *extending* LabelLookup, a lot of the renames and changed declaratiosn would not be necessary, and the patch would only touch half as many files. Also, a lot of the code could keep using a more narrow interface.

Torn between a +1 and a -1 here :)

Patch-set: 2
Label: Code-Review=0