Update patch set 4
Patch Set 4:

EntityView::getHtmlForTermBox uses $entityRevision->getRevisionId() to (in theory) make sure the term box content is consistent with the (cached) content of the view. If the revision ID is stale, this causes an old revision to be used for constructing the term box. Putting 0 here causes the latest version to be used for constructing the term box. 

That's probably better than stale, but may still be inconsistent, if the cached page content isn't the latest, e.g. due to job queue lag, or because the page hasn't yet been re-rendered in a particular language...

Putting 0 here should be find if the parser-cached html always corresponds to the latest revision, as it should... still, feels rather scary. Do we have no way to know which revision of an entity we are currently rendering? We loaded it at some point, right?

Patch-set: 4