Update patch set 16
Patch Set 16: Code-Review+2

I'll merge this, since it'S consistent with the current behavior of the API and UI: dates tagged as julian will be saved as-is.

Note however that the data in the database is a wild mix: some of thje dates tagged as julian are julian, some are gregorian. So some of the resulting dates will be wrong, no matter what we do.

Patch-set: 16
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2