Update patch set 1
Patch Set 1:

After talking with Markus, I am convinced we need to drop the version, for the following reason:

We can change ontology in two ways - add stuff (types, predicates, etc.) or remove stuff. If we want, for example, to change what rank is, we'd add wikibase:NewRank and it would not be the same as wikibase:Rank. It is better than changing what wikibase: means and breaking all types and predicates and classes in the ontology altogether. I.e. we would never do what XSD did - call two things with the same name xsd:dateTime, instead we would do something like xsd:dateTimeISO. 

That said, probablility that we'd change the meaning of some existing name is pretty low. Introducing new names and retiring all names does not break the ontology. So I think we can do without version in the same. Of course, we would need also a link to the current ontology definition saying what the names mean, as discussed in T94747.

Patch-set: 1
Label: Verified=0