Remove broken unlocalization from MwTimeIsoFormatter
This depends on https://github.com/DataValues/Time/pull/62

The formatter applied his steps in a somewhat mixed way:
1. Format the year as a localized number.
2. Put the year in a unlocalized message.
3. Unlocalize the year by applying string replacements. This also
replaced characters that are part of the message.

This patch is only a step forward. I'm removing the localization
for now. It should be re-added later, most probably in an other
place.

The other changes must be in the same patch because of the roundtrip
tests.

Bug: T95532
Change-Id: Ic5b4cd42e5e5690d4787235ab090961d9864989f