Update patch set 4
Patch Set 4:

> we'd have to check all code that accesses the revision ID from an EntityRevision object.

No, there is only a single relevant code path. I checked it. It makes sense to me. I can show you.

> Putting 0 here causes the latest version to be used for constructing the term box. That's probably better than stale, but may still be inconsistent [...]

I agree. But this is out of the scope of this patch. Didn't we discussed this recently? Is a patch ok if it makes things *better* or must it be *perfect*?

I'm going to merge this today if there is no -2.

Patch-set: 4