Update patch set 5
Patch Set 5:

think i have a solution to assign the revision id to wikibase-view-chunks later in the save process once it is assigned.

the solution might not be pretty and it's WIP, but will post that patch soon and doesn't eliminate the need for this patch.

setting it to 0 here though will be needed. In EntityRevision, it means "0 for none" per documentation in the EntityRevision in constructor.  that is indeed the case and is appropriate here.

Patch-set: 5