Update patch set 5
Patch Set 5:

@thiemo: my problem is that I see 43 places in our code calling EntityRevision::getRevisionId. I don't know how many of them rely on the revision to no be 0.

The constructor of EntityRevision sais the revision id can be 0. getRevisionId() *doesn't* say this, and I suspect many callers do not handle 0 nicely. 

Even if *currently* there is only one relevant code path, we'd need a mechanism that would ensure that it will stay that way, and that we will never hit a case where the revision ID is 0, but expected to be non-0.

One easy way would be to have getRevisionId throw an exception if the revision id is < 1. That way, we'd be sure that if it was requested, and there was no exception, it's a "real" revision id.

Patch-set: 5