Update revision id for termbox placeholder during doEditUpdates
This requires a core change to add a hook in the right place,
during doEditUpdates (when we then know the revision id), but
before anything is saved in ParserCache.

Furthermore, the stuff involved here is super nasty but don't
see a way to avoid this. Added @fixme in the code and we
should address it best we can.

If this approach is sane, I would like to then factor out some
of the wikibase-view-chunks hooks code into another class and
add test coverage for it.

Change-Id: Id082e563ef8591fcbe7318d9926d725047b93185