Update patch set 5
Patch Set 5:

> I see 43 places in our code calling EntityRevision::getRevisionId.

Does it help if I repeat that there is only 1 (!) code path relevant for this patch?

> I don't know how many of them rely on the revision to no be 0.

Simple: none. EntityRevision *explicitely* allows the id to be 0. It *always* did that. This is *not* introduced by this patch. Code that relies on it to *not* be 0 does it wrong. I really don't understand why we are discussing this here. This is *completely* unrelated to this patch. Please open a ticket. Also see I50b3f63.

Patch-set: 5