Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

see comment and I'm not sure it's consistently adding labels for properties:

P707 is "sex or gender" for me, and it's mentioned in several summaries but only has a label once. 

I am viewing this in english. This and the other properties here definitely have labels for english.

http://snag.gy/KuRWZ.jpg

someone or myself needs to investigate more why this is not working right, or maybe it's just me.

Patch-set: 2
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1