Update patch set 5
Patch Set 5:

@thiemo: The problem here is that EntityRevision's behavior wrt revision=0 is under-specified. This is my fault: I did not make it clear in the documentation of getRevisionId whether it can return 0. It currently looks like it can't. But the constructor sais you can supply 0. That's inconsistent.

As far as I can tell, EntityRevision is currently always constructed with an actual revision ID. And that revision ID is used in a lot of places, probably not all of which would handle 0 correctly. 

So, I'd rather not start providing 0 there, I'd rather like to disallow 0. In the constructor, or at least make the getter fail.

The thing is... for the one use case relevant to this patch, why do we pass an EntityRevision at all, if it has no revision info? Why do the places we pass it to ask for an EntityRevision, if they do not use the revision info? And if they don't use it now, what keeps them from using it in the future, assuming that the revision ID will be there, and meaningful?

Patch-set: 5