Update patch set 5
Patch Set 5:

> EntityRevision's behavior wrt revision=0 is under-specified.

"Revision ID or 0 for none" in the documentation *and* a default of "$revisionId = 0" *and* a check for "$revisionId < 0". It can't be more specific. The class is so insanely trivial, refering to a "missing" documentation in the getter can't be an argument. The constructor is literally on the same screen.

> It currently looks like it can't.

There is *no* documentation in the getter. How can *no* documentation be confused with "can not be 0"? The getter just doesn't repeat the specification from the constructor, which is consistent to several thousand getters in our code base.

All questions raised here are not relevant for the fix. The class *explicitely* allows 0. This patch uses it as documented.

Patch-set: 5