Allow the badgeItems setting to be null and not have a default
My reason to touch this is the line added to
OutputPageJsConfigHookHandlerTest in If4e080b. I know the setting does
have a default. It defaults to an empty array. I don't have a problem
with that default. But I don't like it if a setting that's empty by
default anyway makes code crash hard with type errors and such if it's
not set (e.g. in tests).

Change-Id: I2777f27acaaf2c0e072c8e703c38adc0864c1c31