Update patch set 2
Patch Set 2:

(2 comments)

What this patch does is not wrong, but please note that it does not change any behavior in production. wbformatvalue API works as expected with and without the added line. So this patch it's actually unrelated to the linked ticket, merging this does not resolve the ticket.

Patch-set: 2