Update patch set 2
Patch Set 2: Code-Review-1

If this is only removed temporarily to be re-added later, perhaps just comment out the code, or at least put a big TODO there.

In any case, it would be good to have the future patch that re-adds localization up on gerrit before we merge a patch that removed the localization step.

Patch-set: 2