Update patch set 2
Patch Set 2:

> just comment out the code, or at least put a big TODO there.

No. It's broken. In the wrong place. Executed in the wrong order. Needs a non-minor rewrite anyway. This patch fixes an actual bug (which is: the message gets destroyed by the so called "unlocalization"). What you want is unrelated and can't be a reason for a -1.

Are we really discussing that question again? Is a patch allowed to be merged if if makes things *better* or must it be *perfect*? I think I made it clear that I strongly disagree with the later.

Patch-set: 2
Label: Verified=0