Update patch set 1
Patch Set 1:

If an installation doesn't want badges, it just doesn't define the setting. The default is array().

The code should not have extra internal defaults, not should it have "nice" handling of unknown/undefined settings: that would hide errors. Any setting accessed by the code should have a pre-defined default value. If it doesn't, something is wrong, and we should see an excepetion, not hide the issue by doing something "sensible".

Patch-set: 1