Update patch set 5
Patch Set 5:

Not a real lot, we don't have that many properties. Also, we do not always output property entities - we only do that in the following cases:

1. Dump
2. We're asked to output them by option 

In the first case, we need yo output everything anyway. In the second case, we could check options and see if we ever used each predicate, but this would be a lot of work for saving 2-3 triples, does not sound like worthy investment of time. 

We could have it in a separate dump part, but again, we have only 2000 properties, and compared to the rest of the dump, it is negligible.

Patch-set: 5
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0