Show the value on Special:SetLabel etc if none is entered
When only an item id and a language are entered, just show
the label/description or aliases instead of showing a warning.
This way the user sees the value, can check it and remove it
if necessary. It prevents users from accidentally removing terms
and is more intuitive as the user actually empties a text field
to remove a value instead of just submitting an already empty one.

Change-Id: I22d44c897700cb796f45d83a52d9e94cef96b1e7