Changed default sharedCacheKeyPrefix setting to be string
The default now uses the somewhat sane assumption that
client + repo are on the same install.

sharedCacheKeyPrefix is the sort of setting where people
(e.g. those that maintain Wikidata) might like to append a
string to it, such as in their local settings based on some
other factor. Appending a string to a closure does not work.
For Wikidata, the build adds setting overrides for this,
so normally not a problem, but still think a string value
for the setting is better and more sane.

Updated and added detailed comments on how to configure
these settings, with reminders that client settings
override repo settings if both client + repo are installed.

Also move shared cache settings from lib to repo / client,
since we want to get rid of stuff in lib. :)

Change-Id: I59b1ad0541bf6eb14098fe5e86f30e9c01706734