Update patch set 1
Patch Set 1:

We use the same copyright message, but the button's text changes from page to page. On Special:NewItem it is 'Create', on Special:SetSiteLink it is 'Set the site link'. So when the message says 'By clicking "###", etc.' ### must be replaced by the button's text. This is what this patch does.

Patch-set: 1