Update patch set 2
Patch Set 2: Code-Review-2

This seems to be an incomplete fragment of a much larger patch. The 3 special pages that are build on top of this abstract base class do not do a 2 step process, as far as I know. To what "second step" are your referring to? -2 for the moment so it doesn't get accidentally merged.

Patch-set: 2
Label: Code-Review=-2