Add heading before sitelinks section
This patch adds a <h2> heading before the sitelinks section
and makes all site link group <h2> headings <h3> headings.
The styling of these headings looks a bit different but
the semantics are correct this way.

The <h2> section is required for semantic reasons and also
to display a collapsible "Site links" section on mobile wikidata.

Furthermore, the <h2> tag got the class section-heading so that
MobileFrontend will recognize it should be collapsed although
it is not a direct descendant of #content.

Because MobileFrontend won't recognize the heading otherwise,
a <span> is inserted with the id of the heading containing the
actual text.

The <h2> is not shown on desktops but is visible on mobile view.

Bug: T95889
Change-Id: I64d91e2f3b24b6fe03feae51d54ddd02a9c62fa3