Track page-touched via UsageTracker.
This is needed to determien which usage entries are stale if we
can no longer rely on all usage entries to be purged at once
during rendering. That would be the case if a page is rendered
in one language, but we would still want to retain usages
tracked for another language.

Change-Id: I6dde64a526a9e321fadd951d2bcac840f461d845