Update patch set 4
Patch Set 4: Code-Review-1

I think I understand what this is supposed to do: avoid removing the sitelink if only an item and a site is given.

But now I can't find a way to remove the sitelink. If I submit the second form with an empty sitelink field, I get an error.

Patch-set: 4