Update patch set 3
Patch Set 3: Code-Review+2

The code of these special pages does some hidden magic: it magically turns into a 2 step process if the field is empty in the 1st step. This patch changes the 2nd step: instead of resubmitting the empty field, the user gets the possibility to either edit (this is new) or empty the field.

There is more magic: the value is set to null. This triggers loading of the original value into the form.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2