Remove debug code from LangLinkHandler
I never see these warnings in the logs.

Only time they might occur is when enabling Wikibase /
sites table somewhere or creating a new wiki.

In those cases, the warnings are not necessary.
For enabling Wikibase, we usually add and populate
the sites table before enabling the extension, and
thus get no warnings.

If adding a new wiki, maybe it's the other way around
but is being worked on and the warnings are not needed.

Change-Id: Ib6ec4f0482e1e7bd4a694ce7867b830d1e8de4fd