Prepare RDF test data for new data-values/time release
This is done in preparation for the comming release 0.7 of
data-values/time. The year in a TimeValue will always be padded to
have 4 digits, but additional leading zeros will be trimmed. By
using test data that already comply to this convention the tests
become (almost) independent of the time component's version.

Bug: T96135
Bug: T96137
Change-Id: I9d2d4c9757a706e2378fa7c78635f189e5ad60f4