Drop unused use clauses from client/lib/repo/view
I have a handy regex that finds use clauses, where the name of the class
is not mentioned in the code below. It seems some team members are using
tools that do not respect class names that are only mentioned in
comments. This can not happen with this regex.

Recheck.

Change-Id: I403ec6908e25ded1342c47e23e723d3405da11f7