Update Wikidata tests for core api changes
as far as I can tell, the actual output for api users
has not changed (yet) and still gets transformed to
use the '*'.

i tried with wbsetsitelink to trigger the badges
error, and with parsevalue and the output is the
same as before the core changes.

The extra '_idx="0"' in the xml format also shouldn't
be a problem, in my opinion.

We still need to remove use of all the deprecated
methods, etc.

Change-Id: I2172c51904213a3876f868babb3d714c31196139