Update patch set 9
Patch Set 9:

i think the change is sensible to do.

I note that https://phabricator.wikimedia.org/F114802 is how the UI currently looks.

https://phabricator.wikimedia.org/F114803 is with the change.

It's slightly odd to not have the edit button inline with the label / heading, and also the amount of spacing between label and description seems somewhat odd (and too much)

Patch-set: 9